Tableau
These Dashboards shows the frequency and duration of flight delays, allowing users to quickly identify which flights are most likely to be delayed and also displays the on-time performance of flights, providing users with a quick overview of which airlines are most reliable in terms of meeting their scheduled departure and arrival times.
The dashboard also includes filters and dropdown menus that allow users to customize their view of the data, such as by airline, airport. This can help users identify trends and patterns in flight performance, as well as potential areas for improvement.